+++ led-display/led-display-model/led-display-model-jpa/src/main/java/me/hmasrafchi/leddisplay/data/jpa/LedRowEntity.java
@@ -3,6 +3,7 @@
  */
 package me.hmasrafchi.leddisplay.data.jpa;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import javax.persistence.ElementCollection;
@@ -22,7 +23,7 @@ import lombok.Data;
 public class LedRowEntity {
 	@Id
 	@GeneratedValue
-	private Integer id;
+	private BigInteger id;
 
 	@OrderColumn
 	@ElementCollection
